Split reducers
