+++ src/redux/reducers/targets.js
@@ -1,23 +1,19 @@
-import { handleActions } from 'redux-actions';
+import {actionCreators} from "../actionCreators";
 
-import { actionCreators } from '../actionCreators';
-
-const initialState = {
-  selectedTarget: null,
-  available: []
-};
-
-export default handleActions(
-  {
-    // Targets Actions
-    [actionCreators.targets.switchTarget]: (state, action) => ({
-      ...state,
-      selectedTarget: action.payload
-    }),
-    [actionCreators.targets.loadSuccess]: (state, action) => ({
-      ...state,
-      available: action.payload
-    })
+export const targetsReducers = {
+  initialState: {
+    selectedTarget: null,
+    available: []
   },
-  initialState
-);
+  reducers: {
+      // Targets Actions
+      [actionCreators.targets.switchTarget]: (state, action) => ({
+        ...state,
+        selectedTarget: action.payload
+      }),
+      [actionCreators.targets.loadSuccess]: (state, action) => ({
+        ...state,
+        available: action.payload
+      })
+    }
+};
